+++ CHANGES.md
@@ -38,6 +38,7 @@
   - Allow runtime resolution of data-driven duration, eg. for `[1 :second]`
   - Update `duration?` to detect a duration dynamically
   - Fix silent wrong parsing by `any->duration` of the EDN form [time unit-keyword]
+  - [BREAKING CHANGE] Add `duration?`, `dur-time` and `dur-unit` fns to the `keypin.type/IDuration` protocol
 - Add `atom?` validator function for Clojure atoms